+++ projects/swimlane/ngx-ui/src/lib/pipes/pipes.module.ts
@@ -5,10 +5,11 @@ import { IterableMapPipe } from './iterable-map.pipe';
 import { FilterPipe } from './filter.pipe';
 import { DecamalizePipe } from './decamelize.pipe';
 import { JSONTreePipe } from './json-tree.pipe';
+import { TimeZonePipe } from './timezone.pipe';
 
 @NgModule({
-  declarations: [IterableMapPipe, FilterPipe, DecamalizePipe, JSONTreePipe],
-  exports: [IterableMapPipe, FilterPipe, DecamalizePipe, JSONTreePipe],
+  declarations: [IterableMapPipe, FilterPipe, DecamalizePipe, JSONTreePipe, TimeZonePipe],
+  exports: [IterableMapPipe, FilterPipe, DecamalizePipe, JSONTreePipe, TimeZonePipe],
   imports: [CommonModule]
 })
 export class PipesModule {}