+++ projects/swimlane/ngx-ui/CHANGELOG.md
@@ -1,6 +1,7 @@
 # CHANGELOG
 
 ## HEAD (Unreleased)
+- Breaking: now requires moment-timezone
 * Feature: Add amTimeZone pipe for timezone display
 * Feature: Support timezone input for ngx-date-time
 * Feature: Support timezone input for ngx-calendar