+++ projects/swimlane/ngx-ui/CHANGELOG.md
@@ -1,12 +1,15 @@
 # CHANGELOG
 
 ## HEAD (Unreleased)
+_(none)_
+
+--------------------
+
+## 24.0.0 (2019-01-14)
 - Breaking: now requires moment-timezone
 * Feature: Add amTimeZone pipe for timezone display
 * Feature: Support timezone input for ngx-date-time
 * Feature: Support timezone input for ngx-calendar
-
---------------------
 
 ## 23.0.0 (2018-12-31)